DRAFT: Добавление модификаторов для предметов (#25)
* Ver 1.5.1 (16.07.25) (#21)

* Zerazun hull upgrade 2 (#20)

* Update Common_Shipyard.c

* Update Common_Shipyard.c

* Update Common_Shipyard.c

* Update Common_Shipyard.c

* Update Common_Shipyard.c

* Update Common_Shipyard.c

* Update Common_Shipyard.c

* Вариант исправления пропажи абордажников на входе в город (#18)

* Enable boarding for hostile trader (#17)

* Поправка учёта квестовых компаньонов (#16)

* Поправка учёта квестовых компаньонов

* Не считаем квестовых компаньонов в лимите офицеров

* Поправка горячих клавиш на экране корабля (#10)

1. Вызываю экран орудий горячей клавишей S.
2. Жму ОК/Отмена в открывшемся окне.
3. Хочу вызвать экран обратно горячей клавишей → его нет. Потому что переменная не обновляется и нужно нажать ещё раз.

Здесь добавил обновление переменной при закрытии окна по ОК/Отмена


* Upd

---------


* ver 1.5.1 (26.07.2025) + Chinese

* ver. 1.5.1 (29.07.25)

* ver. 1.5.1 (31.07.25)

* add modifiers for items

* add modifiers for items 2

* use item modifiers

* callbacks and passive modifiers

* fixes

---------

+++ Program/quests/Sharlie/GoldenGirl.c
@@ -211,13 +211,11 @@ void GoldenGirl_FortFrance(string qName) // на Сен-Пьер
 void GoldenGirl_Message(string qName) // вестовой в Сен-Пьере
 {
 	// проверка на наличие Элен или Мэри
-	if (CheckAttribute(pchar, "questTemp.GoldenGirl_Block"))
-	{
-		SetFunctionTimerCondition("GoldenGirl_FortFrance", 0, 0, 7, false); // таймер
-	}
-	else
+	if (CheckAttribute(pchar, "questTemp.GoldenGirl_Ready") && !CheckAttribute(pchar, "questTemp.GoldenGirl_Block"))
 	{
 		chrDisableReloadToLocation = true;//закрыть локацию
+		DeleteQuestCondition("GoldenGirl_FortFrance");
+		DeleteQuestCondition("GoldenGirl_go");
 		sld = GetCharacter(NPC_GenerateCharacter("GoldenGirl_FraOfficer", "off_fra_1", "man", "man", 25, FRANCE, -1, true, "quest"));
 		FantomMakeCoolFighter(sld, 25, 60, 60, "blade_10", "pistol5", "bullet", 100);
 		DeleteAttribute(sld, "SaveItemsForDead");
@@ -227,6 +225,10 @@ void GoldenGirl_Message(string qName) // вестовой в Сен-Пьере
 		ChangeCharacterAddressGroup(sld, "FortFrance_town", "quest", "quest1");
 		LAi_SetActorType(sld);
 		LAi_ActorDialog(sld, pchar, "", -1, 0);
+	}
+	else
+	{
+		SetFunctionTimerCondition("GoldenGirl_FortFrance", 0, 0, 7, true); // таймер
 	}
 }
